Added Natural Numbers
Number System is Dozenal by default, but can be Hexadecimal if the situation calls for it.
Lower numbers and numbers with lower (and fewer) prime factors are given simpler names with more forward places of articulation.
Nucleus "a" denotes even numbers, "o" numbers divisible by 3, and "i" prime numbers (and "1").
"ao" and "oa" are used for the two multiples of 6 (less "0"), and 0 has "a" even though it "should" be "ao" due to being a very useful number and thus wanting a simple root.
10-13 are partially inspired/derived from their factors.
"Noq" (13) shares an initial consonant with Ni, as they are both one less than a common radix.
Names are chosen to make mishearing one for another difficult.
{ma, ka, QaN} and {po, don} form rough series from front to back places of articulation, emphasizing their importance and relation to eachother.
"oa" is both the digit for "C (x)", as well as the word for "10 (z)".